+++ R/stack_thresh.R
@@ -11,10 +11,6 @@
 #' @param method The thresholding method to use. See
 #'   [autothresholdr::auto_thresh].
 #' @param fail To which value should pixels not exceeeding the threshold be set?
-#' @param skip.consts An array with only one value (a 'constant array') won't
-#'   threshold properly. By default the function would give an error, but by
-#'   setting this parameter to `TRUE`, the array would instead be skipped (the
-#'   function will return the original array) and give a warning.
 #'
 #' @return A 3d array, the thresholded stack. Pillars not exceeding the
 #'   threshold are set to zero. The attribute 'threshold' gives the value used
@@ -26,43 +22,43 @@
 #'                                        package = 'autothresholdr'),
 #'                            as.is = TRUE))
 #' display(normalize(img[, , 1]), method = 'raster')
-#' img_thresh_mask <- MeanStackThresh(img, 'Otsu')
+#' img_thresh_mask <- mean_stack_thresh(img, 'Otsu')
 #' display(img_thresh_mask[, , 1] > 0, method = 'r')
 #' display(normalize(img[, , 1]), method = 'raster')
-#' img_thresh_mask <- MedStackThresh(img, 'Triangle')
+#' img_thresh_mask <- med_stack_thresh(img, 'Triangle')
 #' display(img_thresh_mask[, , 1] > 0, method = 'r')
 #'
 #' @export
-MeanStackThresh <- function(arr3d, method, fail = NA, skip.consts = FALSE) {
+mean_stack_thresh <- function(arr3d, method, fail = NA) {
   stopifnot(length(dim(arr3d)) == 3)
-  if (length(unique(as.vector(arr3d))) == 1 && skip.consts) {
-    warning("Constant array, skipping thresholding.")
-    return(arr3d)
+  if (length(unique(as.vector(arr3d))) == 1) {
+    stop("The array given for thresholding is constant ",
+         "(all the values are the same). Aborting.")
   }
-  thresh <- autothresholdr::auto_thresh(arr3d, method)
-  med.stack <- MeanPillars(arr3d)
-  med.stack.mask <- med.stack > thresh
-  set.indices <- rep(!as.vector(med.stack.mask), dim(arr3d)[3])
+  thresh <- auto_thresh(arr3d, method)
+  mean.stack <- mean_pillars(arr3d)
+  mean.stack.mask <- mean.stack > thresh
+  set.indices <- rep(!as.vector(mean.stack.mask), dim(arr3d)[3])
   arr3d[set.indices] <- fail
   attr(arr3d, "threshold") <- thresh
   arr3d
 }
 
-#' @rdname MeanStackThresh
+#' @rdname mean_stack_thresh
 #'
 #' @examples
-#' img_thresh_mask <- MedStackThresh(img, 'Triangle')
+#' img_thresh_mask <- med_stack_thresh(img, 'Triangle')
 #' display(img_thresh_mask[, , 1] > 0, method = 'r')
 #'
 #' @export
-MedStackThresh <- function(arr3d, method, fail = NA, skip.consts = FALSE) {
+med_stack_thresh <- function(arr3d, method, fail = NA) {
   stopifnot(length(dim(arr3d)) == 3)
-  if (length(unique(as.vector(arr3d))) == 1 && skip.consts) {
-    warning("Constant array, skipping thresholding.")
-    return(arr3d)
+  if (length(unique(as.vector(arr3d))) == 1) {
+    stop("The array given for thresholding is constant ",
+         "(all the values are the same). Aborting.")
   }
-  thresh <- autothresholdr::auto_thresh(arr3d, method)
-  med.stack <- MedianPillars(arr3d)
+  thresh <- auto_thresh(arr3d, method)
+  med.stack <- median_pillars(arr3d)
   med.stack.mask <- med.stack > thresh
   set.indices <- rep(!as.vector(med.stack.mask), dim(arr3d)[3])
   arr3d[set.indices] <- fail