+++ NEWS.md
@@ -1,3 +1,8 @@
+# 0.5.0
+* Rename all exported functions to be in snake_case
+* The skip.consts option in the stack_thresh functions is gone
+  - Now these functions error if you pass them a constant array
+
 # 0.4.0
 * Add MeanStackThresh and MedStackThresh